Portal. Python + Django/app_project/app_project/requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
Django==4.2.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-decouple==3.8
python-telegram-bot==21.0.1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
telegram==0.0.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
zope.interface==6.1
